requirements.txt
discord.py
async-pavlov